+++ git_improved/templates/python_package.py
@@ -1,4 +1,0 @@
-from ..template import Template
-
-class PythonPackageTemplate(Template):
-    pass